requirements.txt
asgiref==3.3.4
black==21.9b0
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.4.0
PyJWT==2.1.0
pytz==2021.1
regex==2021.9.24
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
